Added file to read DE header

+++ lupnt/data/eop.h
@@ -13,7 +13,7 @@ struct EopFileData {
   VecXi years;
   VecXi months;
   VecXi days;
-  VecXi mjds_utc;
+  VecXd mjds_utc;
   VecXd x;
   VecXd y;
   VecXd ut1_utc;
